backend/requirements.txt
alembic==1.12.0
APScheduler==3.10.4
click==8.1.7
Flask==3.0.0
Flask_JWT_Extended==4.5.2
flask_sqlalchemy==3.1.1
Flask_WTF==1.1.1
openai==0.27.9
PyInquirer==1.0.3
pytest==7.4.2
python-dotenv==1.0.0
python_bcrypt==0.3.2
Requests==2.31.0
SQLAlchemy==2.0.20
timeout_decorator==0.5.0
WTForms==3.0.1
pycryptodome==3.19.0
